+++ sktime_dl/deeplearning/cnn/_classifier.py
@@ -4,7 +4,7 @@ import keras
 import numpy as np
 
 from sktime_dl.base.estimators import BaseDeepClassifier
-from sktime_dl.networks.deeplearning import CNNNetwork
+from sktime_dl.deeplearning.cnn._base import CNNNetwork
 
 
 class CNNClassifier(BaseDeepClassifier):
